% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.sdad.demographics-package.R
\docType{package}
\name{dc.sdad.demographics-package}
\alias{dc.sdad.demographics}
\alias{dc.sdad.demographics-package}
\title{dc.sdad.demographics: Social Data Commons: Population Characteristics: Arlington County Civic Association Demographics}
\description{
Allows user to easily get Arlington County Civic Association Demographics data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Eric Oh \email{eo9tv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
